requirements.txt
pyopenssl
toml
certifi
brotli